==> 1_dataprepare.Rmd <==
---
output: html_document
editor_options: 
  chunk_output_type: console
---



```{r,include=FALSE}
library("knitr")
library("dplyr")
library("plyr")
library("dplyr")

opts_knit$set(root.dir = '/Users/lottevanboheemen/Documents/Monash/PhD/Analyses/Data/_3EAA/rawfiles')
```


Reformat the BayEnv2 output by averaging across runs, rescaling using log-transformation and adding loci names.

```{r}

# bayenv2 output files, all populations
A<-read.table("bf_result_all3") 
B<-read.table("bf_result_all2")
C<-read.table("bf_result_all1")

# overage over the 3 independent runs
my.list<-list(A,B,C)
meanmall <-Reduce("+", my.list) / length(my.list)
write.table(meanall,"bf_mean_all.txt",sep="\t",col.names=FALSE,row.names=FALSE)

# extract the Bayes Factors from the dataframe
first <- seq(2,70,3)
  bf <- data.frame(meanmall[,first])
write.table(bf,"bf_mean_all",sep="\t",col.names=FALSE,row.names=FALSE)

# exctract the Spearman's rho from the dateframe
second<-seq(3,70,3)
  rho <- data.frame(meanmall[,second])
  
# exctract the Pearson's R from the dateframe

third<-seq(4,70,3)
  pearr <- data.frame(meanmall[,third])

# log-scale all variables
logA <- log(bf)
logB <- log(B[2:24])
logC <- log(C[2:24])

# bayenv2 output files, North American populations
A<-read.table("bf_result_na3")
B<-read.table("bf_result_na2")
C<-read.table("bf_result_na1")

my.list<-list(A,B,C)
meanm <-Reduce("+", my.list) / length(my.list)
write.table(meanm,"bf_mean_na",sep="\t",col.names=FALSE,row.names=FALSE)

logA <- log(A[2:24])
logB <- log(B[2:24])
logC <- log(C[2:24])


my.listlog<-list(logA,logB,logC)
meanmlna <-Reduce("+", my.listlog) / length(my.listlog)
write.table(meanmlna,"bf_mean_nalog",sep="\t",col.names=FALSE,row.names=FALSE)

# bayenv2 output files,European populations
A<-read.table("bf_result_eu4")
B<-read.table("bf_result_eu5")
C<-read.table("bf_result_eu6")
my.list<-list(A,B,C)
meanm <-Reduce("+", my.list) / length(my.list)

second<-seq(3,70,3)
  rho <- data.frame(meanm[,second])
third<-seq(4,70,3)
  pearr <- data.frame(meanm[,third])

A<-read.table("bf_result_eu3")
B<-read.table("bf_result_eu2")
C<-read.table("bf_result_eu1")

my.list<-list(A,B,C)
meanm <-Reduce("+", my.list) / length(my.list)
write.table(meanm,"bf_mean_eu",sep="\t",col.names=FALSE,row.names=FALSE)

logA <- log(A[2:24])
logB <- log(B[2:24])
logC <- log(C[2:24])


my.listlog<-list(logA,logB,logC)
meanmleu <-Reduce("+", my.listlog) / length(my.listlog)
write.table(meanmleu,"bf_mean_eulog",sep="\t",col.names=FALSE,row.names=FALSE)


# bayenv2 output files, Australian populations
A<-read.table("bf_result_au4")
B<-read.table("bf_result_au5")
C<-read.table("bf_result_au6")
my.list<-list(A,B,C)
meanm <-Reduce("+", my.list) / length(my.list)

A<-read.table("bf_result_au3")
B<-read.table("bf_result_au2")
C<-read.table("bf_result_au1")

my.list<-list(A,B,C)
meanm <-Reduce("+", my.list) / length(my.list)
write.table(meanm,"bf_mean_au",sep="\t",col.names=FALSE,row.names=FALSE)

logA <- log(A[2:24])
logB <- log(B[2:24])
logC <- log(C[2:24])


my.listlog<-list(logA,logB,logC)
meanmlau <-Reduce("+", my.listlog) / length(my.listlog)
write.table(meanmlau,"bf_mean_aulog",sep="\t",col.names=FALSE,row.names=FALSE)

#get loci files from PGD spider which also prepared the input for bayenv2
all.loci<- read.table("all_loci.txt", header = T)

na.loci<- read.table("na.loci")
au.loci<- read.table("au.loci")
eu.loci<- read.table("eu.loci")

```

Add loci names to dataframes with BF

```{r}
bf_all_ddgbs <- cbind(all.loci, bf)
logbf_na_ddgbs <- cbind(na.loci, meanmlna)
logbf_eu_ddgbs <- cbind(eu.loci, meanmleu)
logbf_au_ddgbs <- cbind(au.loci, meanmlau)

write.table(bf_all_ddgbs,"bf_all_ddgbs.txt")

write.table(logbf_na_ddgbs,"logbf_na_ddgbs.txt")
write.table(logbf_eu_ddgbs,"logbf_eu_ddgbs.txt")
write.table(logbf_au_ddgbs,"logbf_au_ddgbs.txt")

```

Repeat above steps for XtX results


```{r}
A<-read.table("xtx_result_all3")
B<-read.table("xtx_result_all2")
C<-read.table("xtx_result_all1")

my.list<-list(A,B,C)
meanmall <-Reduce("+", my.list) / length(my.list)
write.table(meanmall,"xtx_mean_all",sep="\t",col.names=FALSE,row.names=FALSE)


A<-read.table("xtx_result_na3")
B<-read.table("xtx_result_na2")
C<-read.table("xtx_result_na1")

my.list<-list(A,B,C)
meanmna <-Reduce("+", my.list) / length(my.list)
write.table(meanm,"xtx_mean_na",sep="\t",col.names=FALSE,row.names=FALSE)

A<-read.table("xtx_result_eu3")
B<-read.table("xtx_result_eu2")
C<-read.table("xtx_result_eu1")

my.list<-list(A,B,C)
meanmeu <-Reduce("+", my.list) / length(my.list)
write.table(meanm,"xtx_mean_eu",sep="\t",col.names=FALSE,row.names=FALSE)

A<-read.table("xtx_result_au3")
B<-read.table("xtx_result_au2")
C<-read.table("xtx_result_au1")

my.list<-list(A,B,C)
meanmau <-Reduce("+", my.list) / length(my.list)
write.table(meanm,"xtx_mean_au",sep="\t",col.names=FALSE,row.names=FALSE)

meanmall<-meanmall[,2]
meanmna<-meanmna[,2]
meanmeu<-meanmeu[,2]
meanmau<-meanmau[,2]

xtx_all_ddgbs <- cbind(all.loci, meanmall)
xtx_na_ddgbs <- cbind(na.loci, meanmna)
xtx_eu_ddgbs <- cbind(eu.loci, meanmeu)
xtx_au_ddgbs <- cbind(au.loci, meanmau)

write.table(xtx_all_ddgbs,"xtx_all_ddgbs.txt")

write.table(xtx_na_ddgbs,"xtx_na_ddgbs.txt")
write.table(xtx_eu_ddgbs,"xtx_eu_ddgbs.txt")
write.table(xtx_au_ddgbs,"xtx_au_ddgbs.txt")

```

Rho
---
```{r}
A<-read.table("bf_result_all3")
B<-read.table("bf_result_all2")
C<-read.table("bf_result_all1")

my.list<-list(A,B,C)
meanmall <-Reduce("+", my.list) / length(my.list)
write.table(meanall,"bf_mean_all.txt",sep="\t",col.names=FALSE,row.names=FALSE)

first<-seq(2,70,3)
second<-seq(3,70,3)
  rho <- data.frame(meanmall[,second])
third<-seq(4,70,3)
  pearr <- data.frame(meanmall[,third])
rho_all_ddgbs <- cbind(all.loci, rho)

header <- read.table("envheader.txt", header=T)
colnames(rho_all_ddgbs) <- colnames(header)
write.table(rho_all_ddgbs,"rho_allddGBS")


###NA

A<-read.table("bf_result_na4")
B<-read.table("bf_result_na5")
C<-read.table("bf_result_na6")

my.list<-list(A,B,C)
meanmna <-Reduce("+", my.list) / length(my.list)
#write.table(meanna,"bf_mean_na.txt",sep="\t",col.names=FALSE,row.names=FALSE)

first<-seq(2,70,3)
second<-seq(3,70,3)
  rho <- data.frame(meanmna[,second])
third<-seq(4,70,3)
  pearr <- data.frame(meanmna[,third])
rho_na_ddgbs <- cbind(na.loci, rho)

header <- read.table("envheader.txt", header=T)
colnames(rho_na_ddgbs) <- colnames(header)
write.table(rho_na_ddgbs,"rho_naddGBS")

###EU
A<-read.table("bf_result_eu4")
B<-read.table("bf_result_eu5")
C<-read.table("bf_result_eu6")

my.list<-list(A,B,C)
meanmeu <-Reduce("+", my.list) / length(my.list)
#write.table(meaneu,"bf_mean_eu.txt",sep="\t",col.names=FALSE,row.names=FALSE)

first<-seq(2,70,3)
second<-seq(3,70,3)
  rho <- data.frame(meanmeu[,second])
third<-seq(4,70,3)
  pearr <- data.frame(meanmeu[,third])
rho_eu_ddgbs <- cbind(eu.loci, rho)

header <- read.table("envheader.txt", header=T)
colnames(rho_eu_ddgbs) <- colnames(header)
write.table(rho_eu_ddgbs,"rho_euddGBS")

###AU

A<-read.table("bf_result_au4")
B<-read.table("bf_result_au5")
C<-read.table("bf_result_au6")

my.list<-list(A,B,C)
meanmau <-Reduce("+", my.list) / length(my.list)
#write.table(meanau,"bf_mean_au.txt",sep="\t",col.names=FALSE,row.names=FALSE)

first<-seq(2,70,3)
second<-seq(3,70,3)
  rho <- data.frame(meanmau[,second])
third<-seq(4,70,3)
  pearr <- data.frame(meanmau[,third])
rho_au_ddgbs <- cbind(au.loci, rho)

header <- read.table("envheader.txt", header=T)
colnames(rho_au_ddgbs) <- colnames(header)
write.table(rho_au_ddgbs,"rho_auddGBS")
```


###END SCRIPT###
