==> 3a_xtxwind_conv.Rmd <==
---
output: html_document
editor_options: 
  chunk_output_type: console
---
Convergence between outlier windows
===  

> Script by Lotte van Boheemen  

Below code will calculating the number of outliers per sliding window as follows:  
#- Exlude windows without outlier SNPs (Yeaman & Hodgins et al. 2016)
- Calculate the expected frequency of outlier SNPs per number of SNPs in a window:  p = ∑_i(a_i/n_i) (for each range) (Yeaman & Hodgins et al. 2016)
- Calculate the upper 90% CI of the binomial distribution (for each range)
- Check which windows fall 90% CI of the binomial distribution (for each range): top-candidate-outlier windows
- Check overlap in top-candidate-outlier windows between ranges

---  


```{r, include=F}
library("tidyverse")
library(binom)
```


Read in data
```{r}
results_all99 <- read.table("xtxna_1000-1000_top99_50miss_allwinds")

results_na99 <- read.table("xtxna_1000-1000_top99_50miss_allwinds")
results_eu99 <- read.table("xtxeu_1000-1000_top99_50miss_allwinds")
results_au99 <- read.table("xtxau_1000-1000_top99_50miss_allwinds")
```

---   

```{r}
topcan.binom <- function(dat, output){

  sdat <- dat
  set.seed(0)
  nsims <- 10000
  maxn <- max(sdat$num.SNPs)
  n <- seq(0, maxn, by = 1)

  #Calculate the expected frequency of outlier SNPs per number of SNPs in a window:  p = ∑_i(a_i/n_i)  
  totsnp1 <- sum(sdat$num.SNPs)
  totout1 <- sum(sdat$num.outlier)
  expect1 <- (totout1 / totsnp1)

    # calculate quantiles
  sdat$p3 <- qbinom (0.999, sdat$num.SNPs, expect1)

  # Check which windows fall above the expected quantile distribution (for each range): top-candidate-outlier windows
  
  sdat$topcan_q <- (sdat$p3 <= sdat$num.outlier) & (sdat$num.outlier > 1) # change 4-9-16: only call topcan when >1 outlier
  
  coverage <- matrix(NA, nrow=length(n), 1)
  q.upper <- matrix(NA, 1, nrow=length(n))
   
  # simulate y data and 99% quantile
  for(i in 1:length(n)){
    y <- rbinom(nsims, n[i], expect1)
    ul <- qbinom(.999, n[i], expect1)
    q.upper[i,] <- ul

    # move the data into a workable dataframe
    for(j in 1:1){
      sig <- length(y[y/n[i] <= ul[j]])
      coverage[i,j] <- sig/nsims
    }
  }

  write.table(sdat, file = output)
  return(subset(sdat, topcan_q == "TRUE"))
  
}
```

```{r}
topcan.binom(results_all99, "sres1000-1000_all99_50miss_allwinds")

topcan.binom(results_na99, "sres1000-1000_na99_50miss_allwinds")
topcan.binom(results_eu99, "sres1000-1000_eu99_50miss_allwinds")
topcan.binom(results_au99, "sres1000-1000_au99_50miss_allwinds")
```


===

**Overlap in top-candidate outlier windows between ranges**

```{r}
tc_all99 <- read.table("sres1000-1000_all99_50miss_allwinds")
tc_all99 <- subset(tc_all99, topcan_q == "TRUE")

tc_na99 <- read.table("sres1000-1000_na99_50miss_allwinds")
tc_na99 <- subset(tc_na99, topcan_q == "TRUE")
tc_na99$sub1 <- paste(tc_na99$contig, tc_na99$win.start, sep="__")
tc_eu99 <- read.table("sres1000-1000_eu99_50miss_allwinds")
tc_eu99 <- subset(tc_eu99, topcan_q == "TRUE")
tc_eu99$sub1 <- paste(tc_eu99$contig, tc_eu99$win.start, sep="__")
tc_au99 <- read.table("sres1000-1000_au99_50miss_allwinds")
tc_au99 <- subset(tc_au99, topcan_q == "TRUE")
tc_au99$sub1 <- paste(tc_au99$contig,tc_au99$win.start, sep="__")

sres99_naeu <- inner_join(tc_na99, tc_eu99, by="sub1")
write.table(sres99_naeu, "sres99_naeu_1000_50miss_allwinds.txt")
sres99_naau <- inner_join(tc_na99, tc_au99, by="sub1")
write.table(sres99_naau, "sres99_naau_1000_50miss_allwinds.txt")
sres99_euau <- inner_join(tc_eu99, tc_au99, by="sub1")
write.table(sres99_euau ,"sres99_euau_1000_50miss_allwinds.txt")
```

---  

**References**  

Renaut, Sebastien, Gregory L. Owens, and Loren H. Rieseberg. "Shared selective pressure and local genomic landscape lead to repeatable patterns of genomic divergence in sunflowers." Molecular ecology 23.2 (2014): 311-324.

Sam Yeaman, Kathryn A. Hodgins, Katie E. Lotterhos, Haktan Sure5, Simon Nadeau, Jon C. Degner, Kristin A. Nurkowski, Pia Smets, Tongli Wang, Laura K. Gray, Katharina J. Liepe, Andreas Hamann, Jason A. Holliday, Michael C. Whitlock, Loren H. Rieseberg, Sally N. Aitken. "Convergent local adaptation to climate in distantly related conifers". Science 2016


##END SCRIPT##
