==> 6_overlap_gctabfxtx.Rmd <==
---
output: html_document
editor_options: 
  chunk_output_type: console
---



```{r}
library("tidyverse")
library("reshape")#melt
library("car") #Anova
library(gdata)
library(VennDiagram)
library(ggrepel)
```

Calculate probability of overlap
---
```{r}
bfn_all = read.table("bf_90miss_99bf3/mbf_1000_993_binnullW_90miss.txt", header=T)
xn_all = read.table("xtx_1000_99_allwind/mxtx_1000_99_binnullW.txt", header=T) %>% 
  mutate(Range = tolower(Range))

gcta_allwind = read.table("gcta_1000_allwind/gcta_all_1000-1000_top99_allwinds", header=T)
gcta_allwind$windID = paste0(gcta_allwind$contig,"__",gcta_allwind$win.start)

gcta = read.table("gcta_1000_allwind/gcta_1000_top99_supwind", header=T)
gcta$windID = paste0(gcta$contig,"__",gcta$win.start)
```


Subset data with either binom or nullw outliers

```{r}
bfna <- bfn_all%>% 
  filter(Range=="nat") %>% 
  droplevels()

bfeu <- bfn_all%>% 
  filter(Range=="eu") %>% 
  droplevels()

bfau <- bfn_all%>% 
  filter(Range=="au") %>% 
  droplevels()

bfn_sup_NA <- bfn_all%>% 
  filter(binnullw=="yes"& Range=="nat") %>% 
  droplevels()

bfn_sup_EU <- bfn_all%>% 
  filter(binnullw=="yes"& Range=="eu") %>% 
  droplevels()

bfn_sup_AU <- bfn_all%>% 
  filter(binnullw=="yes"& Range=="au") %>% 
  droplevels() 

xn_sup_NA <- xn_all %>% 
  filter(binnullw=="yes"& Range=="nat") %>% 
  droplevels()

xn_sup_EU <- xn_all %>% 
  filter(binnullw=="yes"& Range=="eu") %>%
  droplevels()

xn_sup_AU <-xn_all %>% 
  filter(binnullw=="yes"& Range=="au")%>% 
  droplevels() 
```


Window has to have hit in both XtX AND EAA within a range (based on binom or nullw)

```{r}
bfn_sup_NAall = as.data.frame(levels(as.factor(bfn_sup_NA$windID)))
colnames(bfn_sup_NAall) = "windID"

bfxtx_sup_NA = inner_join(xn_sup_NA, bfn_sup_NAall, by="windID")

bfn_sup_EUall = as.data.frame(levels(as.factor(bfn_sup_EU$windID)))
colnames(bfn_sup_EUall) = "windID"

bfxtx_sup_EU = inner_join(xn_sup_EU, bfn_sup_EUall, by="windID")

bfn_sup_AUall = as.data.frame(levels(as.factor(bfn_sup_AU$windID)))
colnames(bfn_sup_AUall) = "windID"

bfxtx_sup_AU = inner_join(xn_sup_AU, bfn_sup_AUall, by="windID")

bfxtxgcta_sup_NA = select(inner_join(bfxtx_sup_NA, gcta, by="windID"), windID)
bfxtxgcta_sup_EU = select(inner_join(bfxtx_sup_EU, gcta, by="windID"), windID)
bfxtxgcta_sup_AU = select(inner_join(bfxtx_sup_AU, gcta, by="windID"), windID)
```

Test
---

```{r}
#EAA-XTX
phyper(length(intersect(bfxtx_sup_NA$windID,bfxtx_sup_EU$windID))-1, length(levels(as.factor(bfxtx_sup_NA$windID))),  length(intersect(bfna$windID,bfeu$windID)) - length(levels(as.factor(bfxtx_sup_NA$windID))), length(levels(as.factor(bfxtx_sup_EU$windID))), lower.tail = F, log.p = FALSE) #bfxtx_NAEU

phyper(length(intersect(bfxtx_sup_EU$windID,bfxtx_sup_AU$windID))-1, length(levels(as.factor(bfxtx_sup_EU$windID))),  length(intersect(bfeu$windID,bfau$windID)) - length(levels(as.factor(bfxtx_sup_EU$windID))), length(levels(as.factor(bfxtx_sup_AU$windID))), lower.tail = F, log.p = FALSE) #bfxtx_EUAU

phyper(length(intersect(bfxtx_sup_NA$windID,bfxtx_sup_AU$windID))-1, length(levels(as.factor(bfxtx_sup_NA$windID))),  length(intersect(bfna$windID,bfau$windID)) - length(levels(as.factor(bfxtx_sup_NA$windID))), length(levels(as.factor(bfxtx_sup_AU$windID))), lower.tail = F, log.p = FALSE) #bfxtx_NAAU

```

Plot
```{r}

#eaa-xtx
area1 = length(levels(as.factor(bfxtx_sup_NA$windID)))
area2= length(levels(as.factor(bfxtx_sup_EU$windID)))
area3= length(levels(as.factor(bfxtx_sup_AU$windID)))
n12 = length(intersect(bfxtx_sup_NA$windID,bfxtx_sup_EU$windID))
n23 = length(intersect(bfxtx_sup_EU$windID,bfxtx_sup_AU$windID))
n13 = length(intersect(bfxtx_sup_NA$windID,bfxtx_sup_AU$windID))
n123 = length(intersect(intersect(bfxtx_sup_NA$windID,bfxtx_sup_EU$windID), bfxtx_sup_AU$windID))
pdf("venn_eaaxtx.pdf")
draw.triple.venn(area1, area2, area3, n12, n23, n13, n123, col = "transparent", fill = c("cornflowerblue","green","red"),
alpha = 0.50, cex = 1.5, fontface = "bold",cat.cex = 1.5,
cat.pos = 0, cat.dist = 0.07,margin = 0.2)
dev.off()

#all
area1 = length(levels(as.factor(bfxtxgcta_sup_NA$windID)))
area2= length(levels(as.factor(bfxtxgcta_sup_EU$windID)))
area3= length(levels(as.factor(bfxtxgcta_sup_AU$windID)))
n12 = length(intersect(bfxtxgcta_sup_NA$windID,bfxtxgcta_sup_EU$windID))
n23 = length(intersect(bfxtxgcta_sup_EU$windID,bfxtxgcta_sup_AU$windID))
n13 = length(intersect(bfxtxgcta_sup_NA$windID,bfxtxgcta_sup_AU$windID))
n123 = length(intersect(intersect(bfxtxgcta_sup_NA$windID,bfxtxgcta_sup_EU$windID), bfxtxgcta_sup_AU$windID))
draw.triple.venn(area1, area2, area3, n12, n23, n13, n123, col = "transparent", fill = c("cornflowerblue","green","red"),
alpha = 0.50, cex = 1.5, fontface = "bold",cat.cex = 1.5,
cat.pos = 0, cat.dist = 0.07,margin = 0.2)


```


PEA env

```{r}
PEAenv <- c("Longitude", "Latitude", "bio1", "bio2","bio5", "bio10", "bio11")
bfn_PEA <- bfn_all %>% 
  filter(env %in% PEAenv)

```


Subset data with either binom or nullw outliers

```{r}
bfna <- bfn_PEA%>% 
  filter(Range=="nat") %>% 
  droplevels()

bfeu <- bfn_PEA%>% 
  filter(Range=="eu") %>% 
  droplevels()

bfau <- bfn_PEA%>% 
  filter(Range=="au") %>% 
  droplevels()

bfn_sup_NA <- bfn_PEA%>% 
  filter(binnullw=="yes"& Range=="nat") %>% 
  droplevels()

bfn_sup_EU <- bfn_PEA%>% 
  filter(binnullw=="yes"& Range=="eu") %>% 
  droplevels()

bfn_sup_AU <- bfn_PEA%>% 
  filter(binnullw=="yes"& Range=="au") %>% 
  droplevels() 

xn_sup_NA <- xn_all %>% 
  filter(binnullw=="yes"& Range=="nat") %>% 
  droplevels()

xn_sup_EU <- xn_all %>% 
  filter(binnullw=="yes"& Range=="eu") %>%
  droplevels()

xn_sup_AU <-xn_all %>% 
  filter(binnullw=="yes"& Range=="au")%>% 
  droplevels() 
```


Window has to have hit in both XtX AND EAA within a range (based on binom or nullw)

```{r}
bfn_sup_NAall = as.data.frame(levels(as.factor(bfn_sup_NA$windID)))
colnames(bfn_sup_NAall) = "windID"

bfxtx_sup_NA = inner_join(xn_sup_NA, bfn_sup_NAall, by="windID")

bfn_sup_EUall = as.data.frame(levels(as.factor(bfn_sup_EU$windID)))
colnames(bfn_sup_EUall) = "windID"

bfxtx_sup_EU = inner_join(xn_sup_EU, bfn_sup_EUall, by="windID")

bfn_sup_AUall = as.data.frame(levels(as.factor(bfn_sup_AU$windID)))
colnames(bfn_sup_AUall) = "windID"

bfxtx_sup_AU = inner_join(xn_sup_AU, bfn_sup_AUall, by="windID")

bfxtxgcta_sup_NA = select(inner_join(bfxtx_sup_NA, gcta, by="windID"), windID)
bfxtxgcta_sup_EU = select(inner_join(bfxtx_sup_EU, gcta, by="windID"), windID)
bfxtxgcta_sup_AU = select(inner_join(bfxtx_sup_AU, gcta, by="windID"), windID)
```

Test
---

```{r}
#EAA-XTX
phyper(length(intersect(bfxtx_sup_NA$windID,bfxtx_sup_EU$windID))-1, length(levels(as.factor(bfxtx_sup_NA$windID))),  length(intersect(bfna$windID,bfeu$windID)) - length(levels(as.factor(bfxtx_sup_NA$windID))), length(levels(as.factor(bfxtx_sup_EU$windID))), lower.tail = F, log.p = FALSE) #bfxtx_NAEU

phyper(length(intersect(bfxtx_sup_EU$windID,bfxtx_sup_AU$windID))-1, length(levels(as.factor(bfxtx_sup_EU$windID))),  length(intersect(bfeu$windID,bfau$windID)) - length(levels(as.factor(bfxtx_sup_EU$windID))), length(levels(as.factor(bfxtx_sup_AU$windID))), lower.tail = F, log.p = FALSE) #bfxtx_EUAU

phyper(length(intersect(bfxtx_sup_NA$windID,bfxtx_sup_AU$windID))-1, length(levels(as.factor(bfxtx_sup_NA$windID))),  length(intersect(bfna$windID,bfau$windID)) - length(levels(as.factor(bfxtx_sup_NA$windID))), length(levels(as.factor(bfxtx_sup_AU$windID))), lower.tail = F, log.p = FALSE) #bfxtx_NAAU

```

Plot
```{r}

#eaa-xtx
area1 = length(levels(as.factor(bfxtx_sup_NA$windID)))
area2= length(levels(as.factor(bfxtx_sup_EU$windID)))
area3= length(levels(as.factor(bfxtx_sup_AU$windID)))
n12 = length(intersect(bfxtx_sup_NA$windID,bfxtx_sup_EU$windID))
n23 = length(intersect(bfxtx_sup_EU$windID,bfxtx_sup_AU$windID))
n13 = length(intersect(bfxtx_sup_NA$windID,bfxtx_sup_AU$windID))
n123 = length(intersect(intersect(bfxtx_sup_NA$windID,bfxtx_sup_EU$windID), bfxtx_sup_AU$windID))
pdf("venn_eaaxtx_PEA.pdf")
draw.triple.venn(area1, area2, area3, n12, n23, n13, n123, col = "transparent", fill = c("cornflowerblue","green","red"),
alpha = 0.50, cex = 1.5, fontface = "bold",cat.cex = 1.5,
cat.pos = 0, cat.dist = 0.07,margin = 0.2)
dev.off()

#all
area1 = length(levels(as.factor(bfxtxgcta_sup_NA$windID)))
area2= length(levels(as.factor(bfxtxgcta_sup_EU$windID)))
area3= length(levels(as.factor(bfxtxgcta_sup_AU$windID)))
n12 = length(intersect(bfxtxgcta_sup_NA$windID,bfxtxgcta_sup_EU$windID))
n23 = length(intersect(bfxtxgcta_sup_EU$windID,bfxtxgcta_sup_AU$windID))
n13 = length(intersect(bfxtxgcta_sup_NA$windID,bfxtxgcta_sup_AU$windID))
n123 = length(intersect(intersect(bfxtxgcta_sup_NA$windID,bfxtxgcta_sup_EU$windID), bfxtxgcta_sup_AU$windID))
draw.triple.venn(area1, area2, area3, n12, n23, n13, n123, col = "transparent", fill = c("cornflowerblue","green","red"),
alpha = 0.50, cex = 1.5, fontface = "bold",cat.cex = 1.5,
cat.pos = 0, cat.dist = 0.07,margin = 0.2)
dev.off()


```

===

Overlap in BF wind outliers for each candidate environmental variable
---
```{r}
wind_na = read.table("bf_90miss_99bf3/sresbfall_na_1000_993_90miss.txt", header=T)
wind_eu = read.table("bf_90miss_99bf3/sresbfall_eu_1000_993_90miss.txt", header=T)
wind_au = read.table("bf_90miss_99bf3/sresbfall_au_1000_993_90miss.txt", header=T)
wind_na$windID = paste(wind_na$contig, "__", wind_na$win.start)
wind_eu$windID = paste(wind_eu$contig, "__", wind_eu$win.start)
wind_au$windID = paste(wind_au$contig, "__", wind_au$win.start)
wind_na$windIDenv = paste(wind_na$windID, "__", wind_na$envvar)
wind_eu$windIDenv = paste(wind_eu$windID, "__", wind_eu$envvar)
wind_au$windIDenv = paste(wind_au$windID, "__", wind_au$envvar)

wind_na$envvar = as.factor(gsub("_all","",wind_na$envvar))
wind_eu$envvar =  as.factor(gsub("_all","",wind_eu$envvar))
wind_au$envvar =  as.factor(gsub("_all","",wind_au$envvar))
```


```{r}
bfn_all = read.table("bf_90miss_99bf3/mbf_1000_993_binnullW_90miss.txt", header=T)

#binomial only
bf_sup_NA = bfn_all %>% filter(qbinom=="yes" & Range=="nat") %>% droplevels()
bf_sup_EU = bfn_all %>% filter(qbinom=="yes" & Range=="eu") %>% droplevels()
bf_sup_AU= bfn_all %>% filter(qbinom=="yes" & Range=="au") %>% droplevels()

#with null-W
bfn_sup_NA = bfn_all %>% filter(binnullw=="yes" & Range=="nat") %>% droplevels()
bfn_sup_EU = bfn_all %>% filter(binnullw=="yes" & Range=="eu") %>% droplevels()
bfn_sup_AU = bfn_all %>% filter(binnullw=="yes" & Range=="au") %>% droplevels()

naeu = inner_join(bfn_sup_NA,bfn_sup_EU, by="windIDenv")
naau = inner_join(bfn_sup_NA,bfn_sup_AU, by="windIDenv")
euau = inner_join(bfn_sup_EU,bfn_sup_AU, by="windIDenv")

cnaeu = as.data.frame(dplyr::count(naeu, vars=env.x))
cnaau= as.data.frame(dplyr::count(naau, vars=env.x))
ceuau= as.data.frame(dplyr::count(euau, vars=env.x))
call = gdata::combine(cnaeu,cnaau,ceuau)
```



**Calculate the upper 99th quantile of the hypergeometric distribution**  
---
```{r}
topcan.hyper <- function(outdat1, outdat2,fulldat1, fulldat2, comp){
  res = data.frame(env=numeric(), overlap=numeric(),  q99=numeric(), p=numeric())

  for(i in 1:length(levels(fulldat1$envvar))){
    
  envirs =   levels(fulldat1$envvar)
  envir = envirs[i]
  #outdat1 = dataframe with outliers in range 1
  #outdat2 = dataframe with outliers in range 2
  #fuldat1 = dataframe with all SNPs or windows in range 1
  #fulldat2 = dataframe with all SNPs or windows in range 2
  #env = environmental variable to be tested
  
  outdat1env = outdat1 %>% filter(env %in% envir) %>% droplevels()
  outdat2env = outdat2 %>% filter(env %in% envir) %>% droplevels()
  fulldat1env = fulldat1 %>% filter(envvar %in% envir) %>% droplevels()
  fulldat2env = fulldat2 %>% filter(envvar %in% envir) %>% droplevels()

 # phyper(q, m, n, k, lower.tail = FALSE, log.p = FALSE), where q=size of outlier overlap-1; m=number of outlier windows range#1;  #n=(total number shared windows-m); k=number of outlier windows range#2.

q = length(intersect(outdat1env$windIDenv,outdat2env$windIDenv ))   -1
m = nrow(outdat1env)
n = length(intersect(fulldat1env$windIDenv,fulldat2env$windIDenv) )-m
k = nrow(outdat2env)

#I guess expected here is again random distribution, but now I am just showing the top part of the graph
#obsexp_euau20 <- as.data.frame(table(bincount$bin)/(0.05^2*nrow(bincount)))
expected = 0.01^2*length(intersect(fulldat1env$windIDenv,fulldat2env$windIDenv) )

#obsexp = q/expected #I can do this, but I feel it is making it more complicated

a = 1-phyper(q,m,n,k)
b = qhyper(0.999,m,n,k)
#q99 = b/expected #see comment above

  res[i,1] = envir
  res[i,2] = q+1
  res[i,3] = b
  res[i,4] = a

}
res$q = p.adjust(res$p, method = 'fdr')
res$comparison = comp
   return(res)

}

```

Run
---
```{r}
# function(outdat1, outdat2,fulldat1, fulldat2, output)
N_E = topcan.hyper(bfn_sup_NA,bfn_sup_EU, wind_na, wind_eu, "N-E")
N_A = topcan.hyper(bfn_sup_NA,bfn_sup_AU, wind_na, wind_au, "N-A")
E_A = topcan.hyper(bfn_sup_EU,bfn_sup_AU, wind_eu, wind_au, "E-A")

call =rbind(N_E, N_A, E_A)
```

```{r}
call = call %>% filter(! env %in% c("dem"))

pdf("env_binomnullw993_rangeoverlap.pdf", height=3.5, width=7)
ggplot(call,aes(env, overlap, fill=comparison))+
  geom_bar(stat="identity",position='dodge') +
  geom_bar(data = call, aes(env, q99, fill = comparison),stat = 'identity',position="dodge", alpha=0, size=.1, color="black")+ theme_bw() + xlab("Environmental variable")+ theme(axis.text.x  = element_text(angle=45, hjust=1,size=8)) +  ylab("Number of EAA outliers")+scale_fill_manual(values=c("#999999", "#E69F00", "#56B4E9"))
dev.off()

call_sub = call %>% filter( env %in% c("bio1","bio2", "bio5", "bio10","bio11", "Latitude", "Longitude"))

pdf("env_binomnullw993_rangeoverlap_sub.pdf", height=3.5, width=3.5)
ggplot(call_sub,aes(env, overlap, fill=comparison))+
  geom_bar(stat="identity",position='dodge') +
  geom_bar(data = call_sub, aes(env, q99, fill = comparison),stat = 'identity',position="dodge", alpha=0, size=.1, color="black")+ theme_bw() + xlab("Environmental variable")+ theme(axis.text.x  = element_text(angle=45, hjust=1,size=8)) +  ylab("Number of EAA outliers")+ scale_fill_manual(values=c("#999999", "#E69F00", "#56B4E9"))+scale_y_continuous(limits=c(0,45))
dev.off()
```

Without nullW
```{r}
# function(outdat1, outdat2,fulldat1, fulldat2, output)
N_E = topcan.hyper(bf_sup_NA,bf_sup_EU, wind_na, wind_eu, "N-E")
N_A = topcan.hyper(bf_sup_NA,bf_sup_AU, wind_na, wind_au, "N-A")
E_A = topcan.hyper(bf_sup_EU,bf_sup_AU, wind_eu, wind_au, "E-A")

call =rbind(N_E, N_A, E_A)
```

```{r}
call = call %>% filter(! env %in% c("dem"))

pdf("env_binom993_rangeoverlap.pdf", height=3.5, width=7)
ggplot(call,aes(env, overlap, fill=comparison))+
  geom_bar(stat="identity",position='dodge') +
  geom_bar(data = call, aes(env, q99, fill = comparison),stat = 'identity',position="dodge", alpha=0, size=.1, color="black")+ theme_bw() + xlab("Environmental variable")+ theme(axis.text.x  = element_text(angle=45, hjust=1,size=8)) +  ylab("Number of EAA top 1% outliers")+scale_fill_manual(values=c("#999999", "#E69F00", "#56B4E9"))
dev.off()

call_sub = call %>% filter( env %in% c("bio1","bio2", "bio5", "bio10","bio11", "Latitude", "Longitude"))

pdf("env_binom993_rangeoverlap_sub.pdf", height=3.5, width=3.5)
ggplot(call_sub,aes(env, overlap, fill=comparison))+
  geom_bar(stat="identity",position='dodge') +
  geom_bar(data = call_sub, aes(env, q99, fill = comparison),stat = 'identity',position="dodge", alpha=0, size=.1, color="black")+ theme_bw() + xlab("Environmental variable")+ theme(axis.text.x  = element_text(angle=45, hjust=1,size=8)) +  ylab("Number of EAA top 1% outliers")+scale_fill_manual(values=c("#999999", "#E69F00", "#56B4E9"))+scale_y_continuous(limits=c(0,45))
dev.off()
```

for windows
---
```{r}
bfn_all = read.table("bf_90miss_99bf3/mbf_1000_993_binnullW_90miss.txt", header=T)

bf_all_sup = bfn_all %>%
  filter(qbinom=="yes") %>% 
  select(windID, env, Range) %>% 
  droplevels()

colnames(outlw2) = c("Environment","NorthAmerica","Europe","Australia")
write.table(outlw2, "bfwindbinom_outliersperrange.txt")
outlw2 = read.table("bfwindbinom_outliersperrange.txt", header=T)

#with null-W
bf_all_sup = bfn_all %>%
  filter(binnullw=="yes") %>% 
  select(windID, env, Range) %>% 
  droplevels()

outlw <- as.data.frame(table(bf_all_sup$env, bf_all_sup$Range))
colnames(outlw) <- c("Environment", "Range", "windows")
outlw2 <- spread(outlw,  Range, windows)

colnames(outlw2) = c("Environment","NorthAmerica","Europe","Australia")
write.table(outlw2, "bfwindnullw_outliersperrange.txt")
outlw2nullw = read.table("bfwindnullw_outliersperrange.txt", header=T)
```

*plot outliers env* S4
---
```{r}
pdf("_Graph/envoutl_binomwin_NAEU.pdf", height=4, width=4)
ggplot(outlw2, aes(x=NorthAmerica, y=Europe, colour=as.factor(outlwNAEU), shape=as.factor(outlwNAEU),label=Code)) + 
  geom_point() +
  theme_bw() +    theme(panel.grid.minor = element_blank(),  legend.position = "none")+ 
  xlab("Number of binomial outlier windows\n in North America (native)") +
  ylab("Number of binomial outlier windows\n in Europe (introduced)") +
  geom_text_repel()
dev.off()

ggplot(outlw2nullw, aes(x=NorthAmerica, y=Europe, colour=as.factor(outlwNAEU), shape=as.factor(outlwNAEU),label=Code)) + 
  geom_point() +
  theme_bw() +    theme(panel.grid.minor = element_blank(),  legend.position = "none")+ 
  xlab("Number of outlier windows\n in North America (native)") +
  ylab("Number of outlier windows\n in Europe (introduced)") +
  geom_text_repel()

pdf("_Graph/envoutl_binomwin_NAAU.pdf", height=4, width=4)
ggplot(outlw2, aes(x=NorthAmerica, y=Australia, colour=as.factor(outlwNAAU), shape=as.factor(outlwNAAU),label=Code)) + 
  geom_point() +
  theme_bw() +    theme(panel.grid.minor = element_blank(),  legend.position = "none")+ 
  xlab("Number of binomial outlier windows\n in North America (native)") +
  ylab("Number of binomial outlier windows\n in Australia (introduced)") +
  geom_text_repel()
dev.off()

ggplot(outlw2nullw, aes(x=NorthAmerica, y=Australia, colour=as.factor(outlwNAAU), shape=as.factor(outlwNAAU),label=Code)) + 
  geom_point() +
  theme_bw() +    theme(panel.grid.minor = element_blank(),  legend.position = "none")+ 
  xlab("Number of outlier windows\n in North America (native)") +
  ylab("Number of outlier windows\n in Australia (introduced)") +
  geom_text_repel()

pdf("_Graph/envoutl_binomwin_EUAU.pdf", height=4, width=4)
ggplot(outlw2, aes(y=Australia, x=Europe, colour=as.factor(outlwEUAU), shape=as.factor(outlwEUAU),label=Code)) + 
  geom_point() +
  theme_bw() +    theme(panel.grid.minor = element_blank(),  legend.position = "none")+ 
  ylab("Number of binomial outlier windows\n in Australia (introduced)") +
  xlab("Number of binomial outlier windows\n in Europe (introduced)") +
  geom_text_repel()
dev.off()

ggplot(outlw2, aes(y=Australia, x=Europe, colour=as.factor(outlwEUAU), shape=as.factor(outlwEUAU),label=Code)) + 
  geom_point() +
  theme_bw() +    theme(panel.grid.minor = element_blank(),  legend.position = "none")+ 
  ylab("Number of outlier windows\n in Australia (introduced)") +
  xlab("Number of outlier windows\n in Europe (introduced)") +
  geom_text_repel()
```


##END SCRIPT###
