==> 11_LD_process_amongwinds.Rmd <==
---
output: html_document
editor_options: 
  chunk_output_type: console
---
LD analyses
===  

> Script by Lotte van Boheemen  

```{r, include=F}
library("dplyr")
library("plyr")
library("perm")
library("visreg")
library("ade4")
library(lme4)
library(nlme)
library(ggplot2)
library(tidyr)
library("phia")
library("car") #for Anova (with capital A)]
library(lmPerm)
library(gdata)
library(igraph)
devtools::install_github("petrikemppainen/LDna")
library(LDna)
library(heatmap3)
```

---  
Read in the data. Note that below LD is calculated within and not among scaffolds
```{r}
LDNA1 <- read.table("LD/LD_na50_ddgbs_random_results_out_new")
LDEU1 <- read.table("LD/LD_eu50_ddgbs_random_results_out_new")
LDAU1 <- read.table("LD/LD_au50_ddgbs_random_results_out_new")

LDNA2= separate(data = LDNA1, col = V1, into = c("Scaf1", "Loc1"), sep = "__")
LDNA= separate(data = LDNA2, col = V2, into = c("Scaf2", "Loc2"), sep = "__")
LDEU2= separate(data = LDEU1, col = V1, into = c("Scaf1", "Loc1"), sep = "__")
LDEU= separate(data = LDEU2, col = V2, into = c("Scaf2", "Loc2"), sep = "__")
LDAU2= separate(data = LDAU1, col = V1, into = c("Scaf1", "Loc1"), sep = "__")
LDAU= separate(data = LDAU2, col = V2, into = c("Scaf2", "Loc2"), sep = "__")

colnames(LDNA) <- c("Scaf1","Loc1","Scaf2","Loc2","Corr")
colnames(LDEU) <- c("Scaf1","Loc1","Scaf2","Loc2","Corr")
colnames(LDAU) <- c("Scaf1","Loc1","Scaf2","Loc2","Corr")

LDNA$Loc1= as.numeric(LDNA$Loc1)
LDNA$Loc2= as.numeric(LDNA$Loc2)
LDEU$Loc1= as.numeric(LDEU$Loc1)
LDEU$Loc2= as.numeric(LDEU$Loc2)
LDAU$Loc1= as.numeric(LDAU$Loc1)
LDAU$Loc2= as.numeric(LDAU$Loc2)
```

```{r}
LDNA$Dist <- abs(LDNA$Loc2-LDNA$Loc1)
LDNA$sub1 <- paste(LDNA$Scaf1,LDNA$Loc1,sep="__")
LDNA$sub2 <- paste(LDNA$Scaf2,LDNA$Loc2,sep="__")

topSNP_na_2m_rand = read.table("LD/topSNP_na_2m50_randSW_bf3", header=T)
topSNP_na_2m_rand$sub2=topSNP_na_2m_rand$sub1
topSNP_na1 = dplyr::select(topSNP_na_2m_rand, sub1)
topSNP_na2 =dplyr::select(topSNP_na_2m_rand, sub2)

notopSNP_na_rand=  read.table("LD/notopSNP_na_2m50_randSW_bf3", header=T)
notopSNP_na_rand$sub2 = notopSNP_na_rand$sub1
notopSNP_na1 =dplyr::select(notopSNP_na_rand, sub1)
notopSNP_na2 =dplyr::select(notopSNP_na_rand, sub2)

top = left_join(topSNP_na1, LDNA, by="sub1")
top2= left_join(topSNP_na2, top, by = "sub2")
top2$type = "top"
top2<-top2[complete.cases(top2), ]
top_na =dplyr::select(top2, sub1, sub2, Corr)

LDNA_sub = LDNA[which(LDNA$Scaf1!=LDNA$Scaf2),] # remove SNPs on the same scaffold to avoid confounding factors

top = left_join(topSNP_na1, LDNA_sub, by="sub1")
top2= left_join(topSNP_na2, top, by = "sub2")
top2$type = "top"
top2<-top2[complete.cases(top2), ]

notop = left_join(notopSNP_na1, LDNA_sub, by="sub1")
notop2 = left_join(notopSNP_na2, notop, by="sub2")
notop2$type = "notop"
notop2<-notop2[complete.cases(notop2), ]
all_na1 = rbind(top2, notop2)
all_na1$range = "N"
set.seed (1245)
indsamp_na <- rowSums (is.na(notop2)) # counts the number of NA per row
which_na <- which (indsamp_na== 0) #dplyr::select rows which have no NA  
subsampled1K_na <- notop2[sample(which_na,10000,replace = F),]
all_na2 = rbind(top2, subsampled1K_na)
all_na2$range = "N"
```

```{r}
LDEU$Dist <- abs(LDEU$Loc2-LDEU$Loc1)
LDEU$sub1 <- paste(LDEU$Scaf1,LDEU$Loc1,sep="__")
LDEU$sub2 <- paste(LDEU$Scaf2,LDEU$Loc2,sep="__")


topSNP_eu_2m_rand = read.table("LD/topSNP_eu_2m50_randSW_bf3", header=T)
topSNP_eu_2m_rand$sub2=topSNP_eu_2m_rand$sub1
topSNP_eu1 =dplyr::select(topSNP_eu_2m_rand, sub1)
topSNP_eu2 =dplyr::select(topSNP_eu_2m_rand, sub2)

notopSNP_eu_rand=  read.table("LD/notopSNP_eu_2m50_randSW_bf3", header=T)
notopSNP_eu_rand$sub2 = notopSNP_eu_rand$sub1
notopSNP_eu1 =dplyr::select(notopSNP_eu_rand, sub1)
notopSNP_eu2 =dplyr::select(notopSNP_eu_rand, sub2)

top = left_join(topSNP_eu1, LDEU, by="sub1")
top2= left_join(topSNP_eu2, top, by = "sub2")
top2$type = "top"
top2<-top2[complete.cases(top2), ]
top_eu =dplyr::select(top2, sub1, sub2, Corr)

LDEU_sub = LDEU[which(LDEU$Scaf1!=LDEU$Scaf2),] # remove SNPs on the same scaffold to avoid confounding factors

top = left_join(topSNP_eu1, LDEU_sub, by="sub1")
top2= left_join(topSNP_eu2, top, by = "sub2")
top2$type = "top"
top2<-top2[complete.cases(top2), ]

notop = left_join(notopSNP_eu1, LDEU_sub, by="sub1")
notop2 = left_join(notopSNP_eu2, notop, by="sub2")
notop2$type = "notop"
notop2<-notop2[complete.cases(notop2), ]
all_eu1 = rbind(top2, notop2)
all_eu1$range = "E"
set.seed (1245)
indsamp_eu <- rowSums (is.na(notop2)) # counts the number of NA per row
which_eu <- which (indsamp_eu== 0) #dplyr::select rows which have no NA  
subsampled1K_eu <- notop2[sample(which_eu,10000,replace = F),]
all_eu2 = rbind(top2, subsampled1K_eu)
all_eu2$range = "E"
```

```{r}
LDAU$Dist <- abs(LDAU$Loc2-LDAU$Loc1)
LDAU$sub1 <- paste(LDAU$Scaf1,LDAU$Loc1,sep="__")
LDAU$sub2 <- paste(LDAU$Scaf2,LDAU$Loc2,sep="__")


topSNP_au_2m_rand = read.table("LD/topSNP_au_2m50_randSW_bf3", header=T)
topSNP_au_2m_rand$sub2=topSNP_au_2m_rand$sub1
topSNP_au1 =dplyr::select(topSNP_au_2m_rand, sub1)
topSNP_au2 =dplyr::select(topSNP_au_2m_rand, sub2)

notopSNP_au_rand=  read.table("LD/notopSNP_au_2m50_randSW_bf3", header=T)
notopSNP_au_rand$sub2 = notopSNP_au_rand$sub1
notopSNP_au1 =dplyr::select(notopSNP_au_rand, sub1)
notopSNP_au2 =dplyr::select(notopSNP_au_rand, sub2)

top = left_join(topSNP_au1, LDAU, by="sub1")
top2= left_join(topSNP_au2, top, by = "sub2")
top2$type = "top"
top2<-top2[complete.cases(top2), ]
top_au =dplyr::select(top2, sub1, sub2, Corr)

LDAU_sub = LDAU[which(LDAU$Scaf1!=LDAU$Scaf2),] # remove SNPs on the same scaffold to avoid confounding factors

top = left_join(topSNP_au1, LDAU_sub, by="sub1")
top2= left_join(topSNP_au2, top, by = "sub2")
top2$type = "top"
top2<-top2[complete.cases(top2), ]

notop = left_join(notopSNP_au1, LDAU_sub, by="sub1")
notop2 = left_join(notopSNP_au2, notop, by="sub2")
notop2$type = "notop"
notop2<-notop2[complete.cases(notop2), ]
all_au1 = rbind(top2, notop2)
all_au1$range = "A"
set.seed (1245)
indsamp_au <- rowSums (is.na(notop2)) # counts the number of NA per row
which_au <- which (indsamp_au== 0) #dplyr::select rows which have no NA  
subsampled1K_au <- notop2[sample(which_au,10000,replace = F),]
all_au2 = rbind(top2, subsampled1K_au)
all_au2$range = "A"
```

```{r}
all1 = rbind(all_na1, all_eu1, all_au1)
all2 = rbind(all_na2, all_eu2, all_au2)
```

```{r}
all1$RangeTreat <- interaction(all1$range, all1$type)

contr <- contrastCoefficients(
  RangeTreat ~ A.notop-E.notop,
  RangeTreat ~ A.notop-N.notop,
  RangeTreat ~ E.notop-N.notop,
  RangeTreat ~ A.top-E.top,
  RangeTreat ~ A.top-N.top,
  RangeTreat ~ E.top-N.top,
  RangeTreat ~ A.notop-A.top,
  RangeTreat ~ E.notop-E.top,
  RangeTreat ~ N.notop-N.top,
  data=all1, normalize = T)
  
colnames(contr$RangeTreat) <- c("F.AU.vs.EU", "F.AU.vs.NA", "F.EU.vs.NA", "T.AU.vs.EU", "T.AU.vs.NA", "T.EU.vs.NA", "AU.F.vs.T", "EU.F.vs.T","NA.F.vs.T")
```


Test differences between windows and ranges
```{r}
all1$range = as.factor(all1$range)
all1$type = as.factor(all1$type)
hist(sqrt(abs(all1$Corr))) # that is actually pretty good...
mod = lm(sqrt(abs(Corr))~ type*range, data=all1)
Anova(mod, type=3, test.statistic="F")

testInteractions(mod, residual = c("range","type"))

testInteractions(mod, pairwise = "range", slope = "type")
interactionMeans(mod, factor = "range", slope = "type")

testInteractions(mod, fixed="range", across="type")

visreg(mod)

mod = lm(sqrt(abs(Corr))~ RangeTreat, data=all1)
Anova(mod, type=3, test.statistic="F")
testInteractions(mod, custom = contr)
tmp <- interactionMeans(mod, custom=contr)
(tmp$`adjusted mean`)^2
plot(interactionMeans(mod, custom = contr))
```

plot
---
```{r}
dummy=all1
mod = lm(sqrt(abs(Corr))~ type*range, data=all1)
fitted <- predict(mod, interval="confidence")
fitted2 <- fitted^2
dummy<-cbind(dummy,fitted2)

a<-dummy[which(dummy$type=="top"),]
b<-dummy[which(dummy$type=="notop"),]

#pdf("_Graph/LD_amongwinds.pdf", height=6, width=7)
ggplot(aes(y=abs(Corr), x=range, fill=as.factor(range), alpha=as.factor(type), linetype=as.factor(type)), data=dummy) + 
  geom_boxplot(outlier.colour = NA) + theme_bw() + guides(fill=F)+ 
  scale_x_discrete(limits = rev(levels(dummy$range))) + 
  theme(  legend.position = "right") +
  scale_shape_manual(values =c(21,22,24)) + scale_alpha_discrete(range=c(1 ,0.5))+
  scale_y_sqrt()
#dev.off()

pdf("_Graph/LD_amongwinds.pdf", height=3, width=3)
ggplot(aes(y=abs(Corr), x=range, fill=as.factor(range), alpha=as.factor(type), linetype=as.factor(type)), data=dummy) + 
  geom_boxplot(outlier.colour = NA) + theme_bw() + guides(fill=F)+ 
  scale_x_discrete(limits = rev(levels(dummy$range))) + 
  theme(  legend.position = "none", panel.grid.minor = element_blank()) +
  scale_shape_manual(values =c(21,22,24)) + scale_alpha_discrete(range=c(0.5 ,1))+
   scale_linetype_manual(values=c(2,1))+
  scale_y_sqrt(limits=c(0,0.5),breaks=c(0.01,0.05,0.1,0.25,0.5))
dev.off()
```


##END SCRIPT
===
